+++ AdventOfCode.Abstractions/ISolution.cs
@@ -8,10 +8,10 @@ public interface ISolution : ISolutionContext
     /// Executes the logic for part A using the prepared input data.
     /// </summary>
     /// <returns>The computed answer for part A.</returns>
-    public int SolveA();
+    public long SolveA();
     /// <summary>
     /// Executes the logic for part B using the prepared input data.
     /// </summary>
     /// <returns>The computed answer for part B.</returns>
-    public int SolveB();
+    public long SolveB();
 }
